Use `unsafePerformIO`
`unsafeDupablePerformIO` can be interrupted at any time without an
exception and without a change to run finalisers. IO that we preform
unsafely is safe, except that it allocates memory using instances of
`ByteArray` and `allocRet` does not seem to guarantee that `allocRet` is
safe to be use in this context and won’t leak memory if interrupted.

To be fair, it probably _is_ safe, since the way to `allocRet` memory
that won’t leak normally is to allocate garbage-collected memory, which
will be reclaimed even if interrupted.

Anyway, this needs more research, so let’s play safe until we are sure
or until we run benchmarks and now for sure that
`unsafeDupablePerformIO` is woth it for us.