Fix stack nix-shell on Linux
I was previously developing the package on macOS, so had a Darwin-only
system dependency listed. Now I switched to Linux and need to remove it
for the build to work.

A proper fix would be to make a custom shell.nix file for Stack with an
`if` for Darwin, but I am too lazy for this :(.