fix: Add type-level constraing on BLAKE2b key size
The key size used in `blake2bWithKey` must be between
`crypto_generichash_KEYBYTES_MIN` and `crypto_generichash_KEYBYTES_MAX`.
We did not have this constraint in types and accepted a bytearray of any
length instead.

Add this constraint to types. This means that the `key` argument of
`blake2bWithKey` now needs to be a SizedByteArray of an acceptable size.
This might be a little trickier for users to get, but such as life, and
I updated the docs to show how to generate a key.