fix: Remove lower bound on BLAKE2b key size constraint
It turns out that Sodium documentation does not match the implementation
and, in reality, there is no lower bound on the allowed size of the key
passed to blake2b. Many other implementations do not have it. Some users
might actually rely on this fact.

So, we remove this constraint.