mkQuoter: Do not attempt incomplete parses
The previous parser used just `many` to parse literals, which means it
was returning multiple parses (as many parses as there are characters
in the literal).

Stick `eof` at the end to make sure we only parse the literal
completely. This allows for simpler code where we are always guaranteed
to have zero or one elements in the resulting list of possible parses.

Also, do not treat the incomplete parse case separately, since the error
message would have been unhelpful anyway.